requirements.txt
aiogram==3.22.0
pdfplumber==0.11.7
pypdf==6.0.0
python-docx==1.2.0
python-dotenv==1.1.1
